A few minor fixes/tweaks I forgot

+++ src/engine/math_util.h
@@ -75,8 +75,8 @@ void vec3f_set_dist_and_angle(Vec3f from, Vec3f to, f32  dist, s16  pitch, s16  
 s32 approach_s32(s32 current, s32 target, s32 inc, s32 dec);
 f32 approach_f32(f32 current, f32 target, f32 inc, f32 dec);
 #ifdef QOL_FIXES
-s32 approach_s64(s32 current, s32 target, s32 inc, s32 dec);
-f32 approach_f64(f32 current, f32 target, f32 inc, f32 dec);
+s32 approach_s64(s64 current, s64 target, s64 inc, s64 dec);
+f32 approach_f64(f64 current, f64 target, f64 inc, f64 dec);
 #endif
 s16 atan2s(f32 y, f32 x);
 f32 atan2f(f32 a, f32 b);
